% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PersonBaseline.r
\name{PersonBaseline}
\alias{PersonBaseline}
\title{For AIBL longitudinal data, create columns 
person: numerical person ID, max(person) returns the number of participants, 
baseline: value 1 for baseline observation of participant
NumberReps: number of replicates for each person
NOTE: data must be in AIBL.ID and Collection order.}
\usage{
PersonBaseline(ids)
}
\arguments{
\item{ids:}{processed IDS data frame with AIBL.ID and other columns}
}
\description{
For AIBL longitudinal data, create columns 
person: numerical person ID, max(person) returns the number of participants, 
baseline: value 1 for baseline observation of participant
NumberReps: number of replicates for each person
NOTE: data must be in AIBL.ID and Collection order.
}
\examples{
ids<- PersonBaseline(ids)
colnames(ids)
ids[1:20, c("AIBL.ID", "Collection", "Age", "person", "baseline", "NumberReps")]
}
\keyword{IDS}
\keyword{process}
